etl/requirements.txt
pandas
mysql-connector-python
requests
python-dotenv
yfinance
numpy
lxml
certifi
